Finalizado parte 1

+++ src/Resultado_FB.java
@@ -2,6 +2,6 @@
 public class Resultado_FB {
 	int [] ruta;
 	int distancia;
-	int ob;
+	long ob;
 
 }
